backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tensorflow==2.15.0
keras==2.15.0
Pillow==10.1.0
numpy==1.26.2
scikit-learn==1.3.2
kagglehub==0.2.5
python-dotenv==1.0.0
pydantic==2.5.2
aiofiles==23.2.1

